src/main/java/holdem/rankings/OnePairRanking.java
package holdem.rankings;

import holdem.enums.CardGroupRanking;
import holdem.models.Card;

import java.util.List;
import java.util.Map;

import static java.util.stream.Collectors.groupingBy;
import static java.util.stream.Collectors.toList;

public class OnePairRanking extends AbstractRanking {
    public RankingResult parse(List<Card> cards) {
        Map<Integer, List<Card>> groupedCards = cards.stream().collect(groupingBy(Card::getNumber, toList()));

        if (groupedCards.size() == 4 && groupedCards.values().stream().anyMatch(value -> value.size() == 2)) {
            return new RankingResult(CardGroupRanking.ONE_PAIR);
        }
        return null;
    }
}
